pyproject.toml
[tool.poetry]
name = "owl-vectores"
version = "0.1.0"
description = ""
authors = ["felipe.manrique <pmanrique001@gmail.com>"]
readme = "README.md"
packages = [{include = "owl_vectores"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
black = "^23.3.0"
pytest = "^7.3.0"
requests = "^2.28.2"
pypdf2 = "^3.0.1"
langchain = "^0.0.139"
pandas = "^2.0.0"
numpy = "^1.24.2"
python-multipart = "^0.0.6"
tiktoken = "^0.3.3"
python-docx = "^0.8.11"
python-dotenv = "^0.19.2"
chardet = "^5.1.0"
redis = "^4.5.4"
scipy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
